fix index

+++ vyzhanov.daniil/P3/CountUniqueLetters.hpp
@@ -5,7 +5,7 @@
 
 namespace vyzhanov
 {
-  size_t countUniqueLetters(const char* string);
+  int countUniqueLetters(const char* string);
 }
 
 #endif
